requirements.txt
instagrapi
python-telegram-bot
python-dotenv
